requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-pubsub==0.42.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.21.1
httplib2==0.13.0
idna==2.8
oauthlib==3.0.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
uritemplate==3.0.0
urllib3==1.25.3
flask==1.0.3